add logic in lib
